Improve layout co-ordinate handling
Add toX/toY helper methods to encapsulate the common scaling operations
required when adding track so the calculation does not have to be performed at
every call site.

Also changes the internal storage of layout co-ordinates to track the 4
corners, rather than just the origin to allow easier validation that added
tracks are within the layout. Tracks will print errors to the log if they go
out of the layout.